Updated match_scouting.html frontend for 2020 game
Updated the front end of the scouting app for 2020. It looks very bad, but should include all of the questions we need for this years match scouting. HTML and CSS should be updated ASAP so as to avoid the current horrid 5-column number input. Backend is not updated, needs updated to take in the new input values.