Added localStorage, autoSave, and database support
Fixed a lot of bugs caused by previous commits. Also changed match_scouting.js a lot to support storing the data locally (for when the user is disconnected from the server) as well as updating this save on every input done by the user. Changed server.py and database.py to take in data from the form in the new format, and this updated the dataviewer respectively. Frontend still needs debug and updates for the new interface.