Added dividers with buttons for ease of use; Changed css of buttons
Dividers with unique ids were added to match_scouting.html around the team and match number, auto, and teleop questions. Javascript was added to matchScouting.js that lets a previous and next button cycle through which dividers are visible. css was added to form.css to make the buttons look better.